grammar: Replace the Antlr parser for CREATE INDEX statements
Replace the Antlr lexer and parser for CREATE INDEX statements a new
lexer and parser generated with flex and bison. This is intended to be
an experimental first step towards replacing the entire Antlr-generated
parser. This commit is a first step towards that aim and I hope the
tests prove this to be a success, so we can proceed replacing Antlr. In
the meantime, this means that we have to maintain two parsers and that
there is more code, more warnings, more sources of errors, and larger
binary sizes in general. However, if we manage to get rid of the Antlr
parser there are a couple of advantages:
- Getting rid of the Antlr runtime library as a dependency.
- Not depending on an old piece of sotware (we are depending on Antlr2
  while Antlr4 is available at the moment. However, migrating to Antlr4
  is as bad as migrating to bison).
- Better handling of expressions in statements. This proved to be a
  consistent source of problems over the last couple of years.
- Somewhat better Unicode support.
- Reentrant code / multithreading support.
- I can finally uninstall Java from my computer.

See issue #741.